backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
huggingface_hub
pillow